Now on JeliotBroadcast if we are clients, when we start a session we are able to set automaticly the program code on the text area for the programs. The server keeps its own program code and it sends it to the clients. Only we have now to try to set the visualization of Jeliot at the same time on the server and the clients.

+++ branches/JeliotDist/src/examples/Spot.java
@@ -37,7 +37,7 @@ public class Spot {
         }        
     } 
     
-    public int getRadius() {
+    public double getRadius() {
         return radius;
     }
